Uptading MarketServices

+++ Market_System/Market_System/Interface/IMarketable.cs
@@ -10,8 +10,8 @@ namespace Market_System.Interface
     public interface IMarketable
     {
         public List<Product> ShowAllProducts();
-        public int AddProduct(string name, int price, int number, string category);
-        public void UpdateProduct(string name, int number, int price, int productId);
+        public int AddProduct(string name, decimal price, int number, string category);
+        public void UpdateProduct(string name, int number, decimal price, int productId);
         public void RemoveProduct(int id);
         public void ShowProductByCategory(string category);
         public void ShowProductByPriceRange(decimal firstprice, decimal endprice);
